ace_mashup_bot/requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
pip==21.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.0
requests-oauthlib==1.3.0
setuptools==47.1.0
six==1.15.0
tweepy==3.10.0
urllib3==1.26.2
Werkzeug==1.0.1
